01-resnets/code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2018.03.07=0
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h9745a5d_0
cuda91=1.0=h4c16780_0
freetype=2.8=hab7d2ae_1
intel-openmp=2018.0.3=0
jpeg=9b=h024ee3a_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=hdf63c60_3
libgfortran-ng=7.2.0=hdf63c60_3
libpng=1.6.34=hb9fc6fc_0
libprotobuf=3.5.2=0
libstdcxx-ng=7.2.0=hdf63c60_3
libtiff=4.0.9=he85c1e1_1
mkl=2018.0.3=1
mkl_fft=1.0.1=py36h3010b51_0
mkl_random=1.0.1=py36h629b387_0
ncurses=6.1=hf484d3e_0
ninja=1.8.2=py36h6bb024c_1
numpy=1.14.5=py36hcd700cb_0
numpy-base=1.14.5=py36hdbf6ddf_0
olefile=0.45.1=py36_0
openssl=1.0.2o=h20670df_0
pillow=5.1.0=py36h3deb7b8_0
pip=10.0.1=py36_0
protobuf=3.5.2=py36_0
pycparser=2.18=py36hf9f622e_1
python=3.6.5=hc3d631a_2
pytorch=0.4.0=py36_cuda9.1.85_cudnn7.1.2_1
readline=7.0=ha6073c6_4
setuptools=39.2.0=py36_0
six=1.11.0=py36h372c433_1
sqlite=3.24.0=h84994c4_0
tensorboardx=1.2=py_0
tk=8.6.7=hc745277_3
torchvision=0.2.1=py36_1
tqdm=4.23.4=py36_0
wheel=0.31.1=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
